Implement Chatbot class and command generation models for Cyber Query AI application
